fix: relocate W8A8 FlashAttention and split tests
